requirements.txt
asgiref==3.7.2
cffi==1.15.1
cryptography==41.0.1
Django==4.2.2
django-rest-knox==4.2.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
psycopg2==2.9.6
pycparser==2.21
PyJWT==1.7.1
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
